requirements.txt
Django>=5.2.0,<5.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-environ==0.11.2
psycopg[binary]>=3.2.10,<3.4
hdbcli==2.23.27
gunicorn==23.0.0
